From Nicolas Boullis:
  1) work around that vcd_loglevel_default problem,
  2) add *.la.ver (for libvcdinfo.la.ver) to lib/.cvsignore,
  3) remove libiso9660.la from MOSTLYCLEANFILES in lib/Makefile.am.
     (Why did I add this file to MOSTLYCLEANFILES?)


Now, some explanations:

Initially, libvcd was distributed as a shared library, and everything
was fine.

Then we changed it not to have libvcd as a shared library. Now, both
vcd-info and libvcdinfo.so are linked with libvcd. This means that both
have copies of the symbols from libvcd (including vcd_loglevel_default).
At runtime, vcd-info and libvcdinfo.so get linked, and the symbols from
libvcd arrear twice in the namespace. Especially, that means there are
two vcd_loglevel_default variables. By chance, the same one is used
everytime, so it works...

With my current versioning, all symbols in libvcdinfo.so are versioned
with VCDINFO_0, including those from libvcd (although those are not
exported). Then vcd-info uses the unversioned vcd_loglevel_default (I
can't tell why) while libvcdinfo.so uses the VCDINFO_0 one. And that
causes the current breakage.

With the new patch the symbols from libvcd that are in libvcdinfo.so are
unversioned, but they are exported. (I could not find a way to declare
them local without versioning them...) That means we are now back to the
previous behaviour where there are two vcd_loglevel_default symbols at
runtime, but it works...


Now, how to fix this properly?

I think a very good thing to do would be not to have variables at all
shared libraries, as I think it can always be made to fail when the
library is linked twice. Moreover, a shared should only export its own
symbols (not symbols from another library), and only its own symbols
should be used from outside...